Check process names
Checks whether the process with the PID that is stored in the pid file
matches our own. This avoids the possibility of killing the wrong thing
with a 'stop', and ensures that the process will not fail to start when
the pidfile is stale.

This behavior is enabled by default, but can be disabled with a new
config option.